refactor module structure and add infael building
