+++ ninpol/__init__.py
@@ -1,1 +1,3 @@
-from .interpolator import Interpolator
+
+from ._interpolator.interpolator import Interpolator
+from ._interpolator.grid import Grid
