+++ ninpol/_interpolator/ninpol_defines.pxd
@@ -1,0 +1,8 @@
+cpdef enum:
+    NINPOL_MAX_POINTS_PER_ELEMENT = 8
+    NINPOL_MAX_FACES_PER_ELEMENT = 6
+    NINPOL_MAX_POINTS_PER_FACE = 4
+    NINPOL_NUM_ELEMENT_TYPES = 8
+    NINPOL_MAX_EDGES_PER_ELEMENT = 12
+
+
